move group picker to top, remove section lines between group / space privacy / invite link ui for better readability
