+++ src/components/dialogs/Share.vue
@@ -15,6 +15,8 @@ import Embed from '@/components/dialogs/Embed.vue'
 import utils from '@/utils.js'
 import ImportExportButton from '@/components/ImportExportButton.vue'
 import consts from '@/consts.js'
+import AddToGroup from '@/components/dialogs/AddToGroup.vue'
+import GroupLabel from '@/components/GroupLabel.vue'
 
 const globalStore = useGlobalStore()
 const userStore = useUserStore()
@@ -52,7 +54,8 @@ const state = reactive({
   isShareInPresentationMode: false,
   childDialogIsVisible: false,
   spaceUsersIsVisible: false,
-  QRCodeIsVisible: false
+  QRCodeIsVisible: false,
+  addToGroupIsVisible: false
 })
 
 const isSecureAppContextIOS = computed(() => consts.isSecureAppContextIOS)
@@ -100,7 +103,8 @@ const dialogIsVisible = computed(() => {
     state.embedIsVisible ||
     state.childDialogIsVisible ||
     state.spaceUsersIsVisible ||
-    state.QRCodeIsVisible
+    state.QRCodeIsVisible ||
+    state.addToGroupIsVisible
   )
 })
 const closeDialogs = () => {
@@ -110,6 +114,7 @@ const closeDialogs = () => {
   state.childDialogIsVisible = false
   state.spaceUsersIsVisible = false
   state.QRCodeIsVisible = false
+  state.addToGroupIsVisible = false
   globalStore.triggerCloseChildDialogs()
 }
 const childDialogIsVisible = (value) => {
@@ -150,6 +155,11 @@ const toggleQRCodeIsVisible = () => {
   closeDialogs()
   state.QRCodeIsVisible = !isVisible
 }
+const toggleAddToGroupIsVisible = () => {
+  const isVisible = state.addToGroupIsVisible
+  closeDialogs()
+  state.addToGroupIsVisible = !isVisible
+}
 
 // users
 
@@ -159,6 +169,51 @@ const toggleSpaceUsersIsVisible = () => {
   closeDialogs()
   state.spaceUsersIsVisible = value
 }
+
+// group
+
+const userGroups = computed(() => groupStore.getCurrentUserGroups)
+const spaceGroup = computed(() => groupStore.getCurrentSpaceGroup)
+const currentUserIsGroupAdmin = (group) => {
+  return groupStore.getGroupUserIsAdmin({
+    userId: userStore.id,
+    groupId: group.id
+  })
+}
+const toggleSpaceGroup = async (group) => {
+  const currentSpace = spaceStore.getSpaceAllState
+  const shouldRemoveSpaceGroup = currentSpace.groupId === group.id
+  if (shouldRemoveSpaceGroup) {
+    await removeSpaceGroup(group)
+  } else {
+    await updateSpaceGroup(group)
+  }
+  // emit('selectGroup', group)
+}
+const updateSpaceGroup = (group) => {
+  const isSpaceCreator = userStore.getUserIsSpaceCreator
+  if (isSpaceCreator) {
+    groupStore.addSpaceToGroup(group)
+  } else {
+    globalStore.addNotification({
+      message: 'Only space creator can assign to group',
+      type: 'danger'
+    })
+  }
+}
+const removeSpaceGroup = (group) => {
+  const isGroupAdmin = currentUserIsGroupAdmin(group)
+  const isSpaceCreator = userStore.getUserIsSpaceCreator
+  if (isGroupAdmin || isSpaceCreator) {
+    groupStore.removeSpaceFromGroup()
+  } else {
+    globalStore.addNotification({
+      message: 'Only space creator, or group admin, can remove from group',
+      type: 'danger'
+    })
+  }
+}
+
 </script>
 
 <template lang="pug">
@@ -167,10 +222,6 @@ dialog.share.wide(v-if="props.visible" :open="props.visible" @click.left.stop="c
     .row.title-row
       p Share
       .row
-        //- .button-wrap
-        //-   button.small-button
-        //-     img.icon.group(src="@/assets/group.svg")
-        //-     span Groups
         //- users
         .button-wrap
           button.small-button(@click.stop="toggleSpaceUsersIsVisible" :class="{active: state.spaceUsersIsVisible}")
@@ -182,56 +233,69 @@ dialog.share.wide(v-if="props.visible" :open="props.visible" @click.left.stop="c
             span RSS
           RssFeeds(:visible="state.rssFeedsIsVisible")
 
-  section(v-if="spaceIsRemote")
-    PrivacyButton(:privacyPickerIsVisible="state.privacyPickerIsVisible" @togglePrivacyPickerIsVisible="togglePrivacyPickerIsVisible" @closeDialogs="closeDialogs")
+  template(v-if="spaceIsRemote")
+    section
+      //- Group Picker
+      .row.button-wrap.group-button
+        button.group-button(title="Add to Group" :class="{active: state.addToGroupIsVisible}" @click.left.prevent.stop="toggleAddToGroupIsVisible" @keydown.stop.enter="toggleAddToGroupIsVisible")
+          img.icon.group(src="@/assets/group.svg")
+          GroupLabel(v-if="spaceGroup" :group="spaceGroup" :showName="true")
+          template(v-else)
+            span Add to Group
+        AddToGroup(:visible="state.addToGroupIsVisible" @selectGroup="toggleSpaceGroup" :groups="userGroups" :selectedGroup="spaceGroup" @closeDialogs="closeDialogs")
 
-    //- Copy URL
-    section.subsection(v-if="!spaceIsPrivate" :class="{'share-url-subsection-member': isSpaceMember}")
-      .row.title-row
-        .segmented-buttons
-          button(@click.left="copySpaceUrl")
-            img.icon.copy(src="@/assets/copy.svg")
-            .badge.badge-in-button.danger.private-copy-badge(v-if="spaceIsPrivate" title="Private spaces can only be viewed by collaborators")
-              img.icon.lock(src="@/assets/lock.svg")
-            span Copy Public URL
-          button(@click.stop="toggleQRCodeIsVisible" :class="{ active: state.QRCodeIsVisible }" title="Scan QR Code")
-            img.icon.qr-code(src="@/assets/qr-code.svg")
-        QRCode(:visible="state.QRCodeIsVisible" :value="spaceUrl")
+      //- Privacy Picker
+      PrivacyButton(:privacyPickerIsVisible="state.privacyPickerIsVisible" @togglePrivacyPickerIsVisible="togglePrivacyPickerIsVisible" @closeDialogs="closeDialogs")
 
-        .row
-          //- presentation mode
-          label.label.small-button.extra-options-button.inline-button(title="Share in Presentation Mode" @mouseup.prevent.stop.left="toggleIsShareInPresentationMode" @touchend.prevent.stop="toggleIsShareInPresentationMode" :class="{active: state.isShareInPresentationMode}")
-            input(type="checkbox" :value="state.isShareInPresentationMode")
-            img.icon(src="@/assets/presentation.svg")
+      //- Copy URL
+      section.subsection(v-if="!spaceIsPrivate" :class="{'share-url-subsection-member': isSpaceMember}")
+        .row.title-row
+          .segmented-buttons
+            button(@click.left="copySpaceUrl")
+              img.icon.copy(src="@/assets/copy.svg")
+              .badge.badge-in-button.danger.private-copy-badge(v-if="spaceIsPrivate" title="Private spaces can only be viewed by collaborators")
+                img.icon.lock(src="@/assets/lock.svg")
+              span Copy Public URL
+            button(@click.stop="toggleQRCodeIsVisible" :class="{ active: state.QRCodeIsVisible }" title="Scan QR Code")
+              img.icon.qr-code(src="@/assets/qr-code.svg")
+          QRCode(:visible="state.QRCodeIsVisible" :value="spaceUrl")
+          .row
+            //- presentation mode
+            label.label.small-button.extra-options-button.inline-button(title="Share in Presentation Mode" @mouseup.prevent.stop.left="toggleIsShareInPresentationMode" @touchend.prevent.stop="toggleIsShareInPresentationMode" :class="{active: state.isShareInPresentationMode}")
+              input(type="checkbox" :value="state.isShareInPresentationMode")
+              img.icon(src="@/assets/presentation.svg")
 
-  //- Invite
-  InviteToSpace(:visible="isSpaceMember && currentUserIsSignedIn" @closeDialogs="closeDialogs" @childDialogIsVisible="childDialogIsVisible" @selectGroup="selectGroup")
-  section(v-if="!spaceIsRemote")
-    p
-      span To share or invite collaborators,
-      span.badge.info you need to Sign Up or In
-      span for your spaces to be synced and accessible anywhere.
-    button(@click.left="triggerSignUpOrInIsVisible") Sign Up or In
+      //- Invite
+      InviteToSpace(:visible="isSpaceMember && currentUserIsSignedIn" @closeDialogs="closeDialogs" @childDialogIsVisible="childDialogIsVisible" @selectGroup="selectGroup")
 
-  //- Import, Export, Embed
-  section.import-export-section
-    .row
-      ImportExportButton(@childDialogIsVisible="childDialogIsVisible")
-      //- Embed
-      .button-wrap
-        button(@click.left.stop="toggleEmbedIsVisible" :class="{ active: state.embedIsVisible }")
-          span Embed
-        Embed(:visible="state.embedIsVisible")
-    details(@toggle="updateDialogHeight")
-      summary Spread the Word
-      section.subsection
-        p I don't have the resources of a VC backed company, so when you tell a friend about Kinopio, or share spaces at work, it really helps.
-        //- p Your voice is the water that grows this seedling.
-        p – Piri
-        p
-          img(src="https://cdn.kinopio.club/fqoJozHGrobicZe0XUG1e/my-garden.webp")
-        //- https://alternativeto.net/software/kinopio/about/
-        //- https://toolfinder.co/tools/kinopio
+    //- Import, Export, Embed
+    section.import-export-section
+      .row
+        ImportExportButton(@childDialogIsVisible="childDialogIsVisible")
+        //- Embed
+        .button-wrap
+          button(@click.left.stop="toggleEmbedIsVisible" :class="{ active: state.embedIsVisible }")
+            span Embed
+          Embed(:visible="state.embedIsVisible")
+      details(@toggle="updateDialogHeight")
+        summary Spread the Word
+        section.subsection
+          p I don't have the resources of a VC backed company, so when you tell a friend about Kinopio, or share spaces at work, it really helps.
+          //- p Your voice is the water that grows this seedling.
+          p – Piri
+          p
+            img(src="https://cdn.kinopio.club/fqoJozHGrobicZe0XUG1e/my-garden.webp")
+          //- https://alternativeto.net/software/kinopio/about/
+          //- https://toolfinder.co/tools/kinopio
+
+  template(v-else)
+    section
+      p
+        span To share or invite collaborators,
+        span.badge.info you need to Sign Up or In
+        span for your spaces to be synced and accessible anywhere.
+      button(@click.left="triggerSignUpOrInIsVisible") Sign Up or In
+
 </template>
 
 <style lang="stylus">
@@ -266,6 +330,8 @@ dialog.share
     margin-top 0
     border-top-left-radius 0
     border-top-right-radius 0
+  // .borderless-section
+  //   border
 
   .segmented-buttons
     z-index 1
@@ -284,6 +350,7 @@ dialog.share
         pointer-events none
 
   .privacy-button
+    margin 0
     width 100%
     button
       width 100%
